計算結果のオブジェクトを作る処理をモジュールに切り出す
またフラッシュメッセージのhotwireで追加
