+++ app/views/layouts/application.html.erb
@@ -12,11 +12,9 @@
   </head>
 
   <body>
-    <div class="notifications">
-      <% flash.each do |key, value| %>
-        <%= content_tag(:div, value, class: key) %>
-      <% end %>
-    </div>
+    <turbo-frame id="flash">
+      <%= render "shared/flash" %>
+    </turbo-frame>
     <%= render "shared/header" %>
     <%= yield %>
     <%= render "shared/footer" %>
