+++ app/views/shared/_flash.html.erb
@@ -1,0 +1,7 @@
+<div class="flash-messages">
+  <% flash.each do |type, message| %>
+    <div class="flash-message">
+      <%= message %>
+    </div>
+  <% end %>
+</div>
